requirements.txt
av==14.1.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
flake8==7.1.2
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pillow-avif-plugin==1.4.6
platformdirs==4.3.6
pyav==14.1.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
